+++ utils/__init__.py
@@ -1,6 +1,5 @@
 from .classes import *
 from .system import *
-from .scenarios import *
 
 from utils.lang.en import *
 from utils.lang.ru import *