+++ utils/lang/en.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from modules.words2num import words2num
 from num2words import num2words
 import requests
 import random